+++ src/main/resources/application.yml
@@ -1,23 +1,21 @@
 spring:
-  profiles:
-    active: dev
 
   jpa:
-    database: MYSQL
+    generate-ddl: true
+    database: mysql
     hibernate:
-      ddl-auto: validate
+      ddl-auto: none
       naming:
         physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
     show-sql: false
     properties:
       hibernate:
-        dialect: org.hibernate.dialect.MySQL57Dialect
+        dialect: org.hibernate.dialect.MySQLDialect
 
   datasource:
-    url: jdbc:mysql://localhost:3307/test_db
-    username: user
-    password: pass
-    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://localhost/my_app
+    username: my_app
+    password: se3ret
 
   flyway:
     baseline-on-migrate: true
