+++ src/test/java/com/example/test/TestApplicationDataTests.java
@@ -1,0 +1,52 @@
+package com.example.test;
+
+import static org.junit.Assert.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.BindMode;
+import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+import com.example.test.customer.CustomerRepository;
+import com.example.test.project.ProjectRepository;
+
+@DataJpaTest
+@AutoConfigureTestDatabase(replace = Replace.NONE)
+@Testcontainers(disabledWithoutDocker = true)
+class TestApplicationDataTests {
+    @Container
+    public static MySQLContainer<?> mySQLContainer = new MySQLContainer<>("mysql:8.0.33")
+            .withFileSystemBind("./mysql_test_container_volume", "/var/lib/mysql", BindMode.READ_WRITE)
+            .withStartupAttempts(1)
+            .withStartupTimeoutSeconds(1200)
+            .withDatabaseName("my_app")
+            .withUsername("my_app");
+
+    @DynamicPropertySource
+    static void registerMySqlProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url", mySQLContainer::getJdbcUrl);
+        registry.add("spring.datasource.password", mySQLContainer::getPassword);
+    }
+
+    @Autowired
+    private ProjectRepository projectRepository;
+    @Autowired
+    private CustomerRepository customerRepository;
+
+    @Test
+    void tester() {
+        assertNotNull(projectRepository);
+        assertNotNull(customerRepository);
+        assertTrue(projectRepository.findAll().isEmpty());
+        assertTrue(customerRepository.findAll().isEmpty());
+    }
+
+}
