+++ src/test/java/com/example/test/TestApplicationTests.java
@@ -1,13 +1,23 @@
 package com.example.test;
 
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+import javax.sql.DataSource;
+
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@SpringBootTest // passes as long "DEV-DB" is running (as specified in
+                // application.yaml/startDevDeb.sh)
+@AutoConfigureTestDatabase(replace = Replace.NONE)
 class TestApplicationTests {
 
     @Test
-    void contextLoads() {
+    void contextLoads(@Autowired DataSource ds) {
+        assertNotNull(ds);
     }
 
 }
