--- src/test/java/com/example/test/TestDbContainerInitializer.java
@@ -1,23 +1,0 @@
-package com.example.test;
-
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.MySQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-@Testcontainers(disabledWithoutDocker = true)
-public interface TestDbContainerInitializer {
-
-    @Container
-    MySQLContainer<TestDbContainer> mySQLContainer = TestDbContainer.getInstance();
-
-    @DynamicPropertySource
-    static void registerMySqlProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", mySQLContainer::getJdbcUrl);
-        registry.add("spring.datasource.username", mySQLContainer::getUsername);
-        registry.add("spring.datasource.password", mySQLContainer::getPassword);
-        registry.add("spring.datasource.driver-class-name",()->"com.mysql.cj.jdbc.Driver");
-        registry.add("spring.test.database.replace",()->"none");
-    }
-}
