--- src/test/resources/application.yml
@@ -1,6 +1,0 @@
-spring:
-  profiles:
-    active: test
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
